+++ support/src/main/java/com/cmeza/spring/jdbc/repository/support/annotations/methods/supports/JdbcMapping.java
@@ -14,4 +14,8 @@ public @interface JdbcMapping {
     String from() default "";
 
     int type();
+
+    boolean nulled() default false;
+
+    int order() default 0;
 }