+++ support/src/main/java/com/cmeza/spring/jdbc/repository/support/definitions/MappingDefinition.java
@@ -9,4 +9,5 @@ public class MappingDefinition {
     private int type;
     private int position;
     private Object value;
+    private boolean nulled;
 }